add tempest test to activate library policies
Test activating library policies in dsvm job.
Helpful for testing contributions to policy library.
Ideally the test would be done in local unit tests,
but with the way policy creation works right now,
a functioning datasource (eg. nova) is required.

It may be helpful in the future to add a new way of
validating library policies without the running datasource.

Change-Id: I36d78a7f2764386a1fc251dedd4ecfa696000ec1