Resolve test_datasource_db_sync_add_remove instability
Configure the replica server in the test to use a short PRC timeout.
The instability appears to arise as follows.
In starting the replica Congress instance:
If Congress API process starts before Congress policy engine process,
the call to check Congress started is processed by the API process,
which makes an RPC call to the policy engine process.
But the policy engine process may be a bit slower to start.
In that case, the RPC call may never reach the policy engine process
even if the policy engine process starts up just a few seconds later.
So the call times out according to the default 60 seconds, by which
point the entire retry period of the test is used up and the test
fails.

In this fix, the RPC timeout is set to a short 10 seconds on the
replica instance. That way, if the policy engine has not started
by the time of the first RPC call, the RPC call would time out with
time remaining on the test retry period for further retry.

The test retry period is also increased because occassionally the
replica policy engine takes longer than the allotted 60s to start up.

Change-Id: Ied8c95c4a60a3b6ea2c617fdfe94d4d7cf7a5822