Move legacy job to bionic and stable on xenial
We are migrating all the legacy jobs to bionic now
- https://review.openstack.org/#/c/639096

Above patch move all the legacy base jobs to bionic which
will automatically move all derived jobs to bionic.

As congress-tempest-plugin is branchless so these jobs running
on stable branch needs to be on xenial and this also has been
taken care in base patch.

stable job running on master gate has to be on xenial.

This commit keep the stable branch master job on xenial
and rest all on bionic.

Depends-On: https://review.openstack.org/#/c/639096
Change-Id: Id4e0312ff5093aae9704e400bfa74803520bc0e4